// src/app/views/app/products/interfaces/products.interfaces.ts
export interface Product {
  _id:                string;
  brand:              string;
  category:           string;
  description:        string;
  discountPercentage: number;
  images:             string[];
  price:              number;
  rating:             number;
  stock:              number;
  thumbnail:          string;
  title:              string;
 }